+++ eppy/geometry/surface.py
@@ -74,7 +74,7 @@ def unit_normal(pt_a, pt_b, pt_c):
     z_val = np.linalg.det(
         [[pt_a[0], pt_a[1], 1], [pt_b[0], pt_b[1], 1], [pt_c[0], pt_c[1], 1]]
     )
-    magnitude = (x_val ** 2 + y_val ** 2 + z_val ** 2) ** 0.5
+    magnitude = (x_val**2 + y_val**2 + z_val**2) ** 0.5
     if magnitude < 0.00000001:
         raise ZeroDivisionError
     mag = (x_val / magnitude, y_val / magnitude, z_val / magnitude)