+++ eppy/modeleditor.py
@@ -29,6 +29,7 @@ from eppy.idfreader import convertafield
 from eppy.idfreader import makeabunch
 from eppy.runner.run_functions import run
 from eppy.runner.run_functions import wrapped_help_text
+from eppy import idfreader
 
 
 class NoObjectError(Exception):
@@ -768,6 +769,23 @@ class IDF(object):
             block=self.block,
             defaultvalues=defaultvalues,
         )
+        
+        # add fields if there are not enough fields in the IDD to match the fields in kwargs 
+        dtls = self.model.dtls
+        key = obj[0].upper()
+        key_i = dtls.index(key)
+        objfields = [comm.get("field") for comm in self.idd_info[key_i]]
+        # check if there are enough fields in the IDD to match the kwargs        
+        if len(kwargs) > (len(objfields) - 1): # objfields has placeholder for key. So subtract 1
+            # -- increase the number of fields in the IDD (in block and commdct)
+            n = len(kwargs) - (len(objfields) - 1) # objfields has placeholder for key. So subtract 1
+            key_txt = key
+            obj_i = key_i
+            block = self.block
+            commdct = self.idd_info
+            objfields = idfreader.increaseIDDfields(block, commdct, obj_i, key_txt, n)
+            
+        
         abunch = obj2bunch(self.model, self.idd_info, obj)
         self.idfobjects[key].append(abunch)
         for k, v in list(kwargs.items()):