+++ eppy/bunch_subclass.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2012, 2020 Santosh Philip
+# Copyright (c) 2012, 2020, 2022 Santosh Philip
 # Copyright (c) 2016 Jamie Bull
 # Copyright (c) 2020 Cheng Cui
 # =======================================================================