+++ eppy/idfreader.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2012 Santosh Philip
+# Copyright (c) 2012, 2022 Santosh Philip
 # Copyright (c) 2021  Dimitris Mantas
-# Copyright (c) 2022 Santosh Philip
 # =======================================================================
 #  Distributed under the MIT License.
 #  (See accompanying file LICENSE or copy at